test: refactor server folder to a single file
