--- cypress/support/server/index.js
@@ -1,1 +1,0 @@
-export { default as loginAndPersistSession } from './loginAndPersistSession.js'
